move aws wrapper structs to use service interfaces
