+++ api/handlers_services.go
@@ -12,12 +12,13 @@ import (
 
 	"github.com/YaleSpinup/ecs-api/apierror"
 	"github.com/YaleSpinup/ecs-api/orchestration"
+	"github.com/pkg/errors"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/arn"
 	"github.com/aws/aws-sdk-go/service/cloudwatchlogs"
 	"github.com/aws/aws-sdk-go/service/ecs"
-	"github.com/aws/aws-sdk-go/service/servicediscovery"
+	"github.com/aws/aws-sdk-go/service/ecs/ecsiface"
 
 	"github.com/gorilla/mux"
 
@@ -157,9 +158,7 @@ func (s *server) ServiceUpdateHandler(w http.ResponseWriter, r *http.Request) {
 
 	var req orchestration.ServiceOrchestrationUpdateInput
 	if err := json.NewDecoder(bytes.NewReader(body)).Decode(&req); err != nil {
-		log.Error("cannot Decode body into update service input")
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte(err.Error()))
+		handleError(w, errors.Wrap(err, "cannot decode body into update service input"))
 		return
 	}
 	log.Debugf("decoded request into service (%s/%s) orchestration request:\n%+v", cluster, service, req)
@@ -172,16 +171,13 @@ func (s *server) ServiceUpdateHandler(w http.ResponseWriter, r *http.Request) {
 
 	output, err := orchestrator.UpdateService(r.Context(), cluster, service, &req)
 	if err != nil {
-		log.Errorf("error in service update orchestration: %s", err)
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte(err.Error()))
+		handleError(w, err)
 		return
 	}
 
 	j, err := json.Marshal(output)
 	if err != nil {
-		log.Errorf("cannot marshal response (%v) into JSON: %s", output, err)
-		w.WriteHeader(http.StatusInternalServerError)
+		handleError(w, errors.Wrap(err, "cannot decode output into update service input"))
 		return
 	}
 
@@ -204,33 +200,15 @@ func (s *server) ServiceListHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Collect all of the task
-	input := ecs.ListServicesInput{
-		Cluster:    aws.String(cluster),
-		LaunchType: aws.String("FARGATE"),
-	}
-	output := []string{}
-	for {
-		out, err := ecsService.Service.ListServicesWithContext(r.Context(), &input)
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte(err.Error()))
-			return
-		}
-
-		for _, t := range out.ServiceArns {
-			output = append(output, aws.StringValue(t))
-		}
-
-		if out.NextToken == nil {
-			break
-		}
-		input.NextToken = out.NextToken
+	output, err := ecsService.ListServices(r.Context(), cluster)
+	if err != nil {
+		handleError(w, err)
+		return
 	}
 
 	j, err := json.Marshal(output)
 	if err != nil {
-		log.Errorf("cannot marshal response (%v) into JSON: %s", output, err)
-		w.WriteHeader(http.StatusInternalServerError)
+		handleError(w, errors.Wrap(err, "unable to marshal response from the ssm service"))
 		return
 	}
 
@@ -262,19 +240,9 @@ func (s *server) ServiceShowHandler(w http.ResponseWriter, r *http.Request) {
 		all = b
 	}
 
-	serviceOutput, err := ecsService.Service.DescribeServicesWithContext(r.Context(), &ecs.DescribeServicesInput{
-		Cluster:  aws.String(cluster),
-		Services: aws.StringSlice([]string{service}),
-	})
+	serviceOutput, err := ecsService.GetService(r.Context(), cluster, service)
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte(err.Error()))
-		return
-	}
-
-	if len(serviceOutput.Services) == 0 {
-		w.WriteHeader(http.StatusNotFound)
-		w.Write([]byte("Not Found"))
+		handleError(w, err)
 		return
 	}
 
@@ -282,25 +250,21 @@ func (s *server) ServiceShowHandler(w http.ResponseWriter, r *http.Request) {
 	if !all {
 		j, err = json.Marshal(serviceOutput)
 		if err != nil {
-			log.Errorf("cannot marshal response (%v) into JSON: %s", serviceOutput, err)
-			w.WriteHeader(http.StatusInternalServerError)
+			handleError(w, err)
 			return
 		}
 	} else {
 		log.Debugf("getting all details about %s/%s", cluster, service)
-		tdOutput, err := ecsService.Service.DescribeTaskDefinitionWithContext(r.Context(), &ecs.DescribeTaskDefinitionInput{
-			TaskDefinition: serviceOutput.Services[0].TaskDefinition,
-		})
+
+		tdOutput, err := ecsService.GetTaskDefinition(r.Context(), serviceOutput.TaskDefinition)
 		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte(err.Error()))
+			handleError(w, err)
 			return
 		}
 
 		tasks, err := tasksList(r.Context(), ecsService.Service, cluster, service)
 		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte(err.Error()))
+			handleError(w, err)
 			return
 		}
 
@@ -308,7 +272,7 @@ func (s *server) ServiceShowHandler(w http.ResponseWriter, r *http.Request) {
 		sd, ok := s.sdServices[account]
 		if ok {
 			log.Debugf("found service discovery account information for all details lookup of %s/%s", cluster, service)
-			serviceDiscoveryEndpoint, err = serviceEndpoint(r.Context(), sd.Service, serviceOutput.Services[0].ServiceRegistries[0])
+			serviceDiscoveryEndpoint, err = sd.ServiceEndpoint(r.Context(), aws.StringValue(serviceOutput.ServiceRegistries[0].RegistryArn))
 			if err != nil {
 				log.Errorf("error getting servicediscovery endpoint for %s/%s: %s", cluster, service, err)
 			}
@@ -320,16 +284,15 @@ func (s *server) ServiceShowHandler(w http.ResponseWriter, r *http.Request) {
 			Tasks           []*string
 			TaskDefinition  *ecs.TaskDefinition
 		}{
-			Service:         serviceOutput.Services[0],
+			Service:         serviceOutput,
 			ServiceEndpoint: serviceDiscoveryEndpoint,
 			Tasks:           tasks,
-			TaskDefinition:  tdOutput.TaskDefinition,
+			TaskDefinition:  tdOutput,
 		}
 
 		j, err = json.Marshal(output)
 		if err != nil {
-			log.Errorf("cannot marshal response (%v) into JSON: %s", output, err)
-			w.WriteHeader(http.StatusInternalServerError)
+			handleError(w, err)
 			return
 		}
 	}
@@ -377,7 +340,7 @@ func (s server) newOrchestrator(account string) (*orchestration.Orchestrator, er
 }
 
 // tasksList collects all of the task ids (with the disired state of both running and stopped) for a service
-func tasksList(ctx context.Context, es *ecs.ECS, cluster, service string) ([]*string, error) {
+func tasksList(ctx context.Context, es ecsiface.ECSAPI, cluster, service string) ([]*string, error) {
 	tasks := []*string{}
 	runningTaskOutput, err := es.ListTasksWithContext(ctx, &ecs.ListTasksInput{
 		Cluster:     aws.String(cluster),
@@ -412,47 +375,6 @@ func tasksList(ctx context.Context, es *ecs.ECS, cluster, service string) ([]*st
 	return tasks, nil
 }
 
-// serviceEndpoint takes the service discovery client and the ecs service registry configuration.  It first gets the
-// details of the service registry given and from that determines the namespace ID.  The endpoint string is determined by
-// combining the service registry service name(hostname) and the namespace name (domain name).
-func serviceEndpoint(ctx context.Context, sd *servicediscovery.ServiceDiscovery, registry *ecs.ServiceRegistry) (*string, error) {
-	serviceResistryArn, err := arn.Parse(aws.StringValue(registry.RegistryArn))
-	if err != nil {
-		log.Errorf("error parsing servicediscovery service ARN %s", err)
-		return nil, err
-	}
-
-	if serviceResistryArn.Resource != "" {
-		log.Debugf("getting service registry service with id %s", serviceResistryArn.Resource)
-
-		// serviceRegistryArn.Resource is of the format 'service/srv-xxxxxxxxxxxxx', but GetServiceInput needs just the ID
-		serviceID := strings.SplitN(serviceResistryArn.Resource, "/", 2)
-		sdOutput, err := sd.GetServiceWithContext(ctx, &servicediscovery.GetServiceInput{
-			Id: aws.String(serviceID[1]),
-		})
-
-		if err != nil {
-			log.Errorf("error getting service from ID %s: %s", serviceID[1], err)
-			return nil, err
-		}
-
-		if nsID := aws.StringValue(sdOutput.Service.DnsConfig.NamespaceId); nsID != "" {
-			namespaceOutput, err := sd.GetNamespaceWithContext(ctx, &servicediscovery.GetNamespaceInput{
-				Id: aws.String(nsID),
-			})
-			if err != nil {
-				log.Errorf("error getting namespace %s", err)
-				return nil, err
-			}
-			endpoint := fmt.Sprintf("%s.%s", aws.StringValue(sdOutput.Service.Name), aws.StringValue(namespaceOutput.Namespace.Name))
-			return &endpoint, nil
-		}
-	}
-
-	log.Warnf("service discovery endpoint not found")
-	return nil, nil
-}
-
 // ServiceEventsHandler gets the events for a service in a cluster
 func (s *server) ServiceEventsHandler(w http.ResponseWriter, r *http.Request) {
 	w = LogWriter{w}
@@ -467,27 +389,16 @@ func (s *server) ServiceEventsHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	output, err := ecsService.Service.DescribeServicesWithContext(r.Context(), &ecs.DescribeServicesInput{
-		Cluster:  aws.String(cluster),
-		Services: aws.StringSlice([]string{service}),
-	})
+	output, err := ecsService.GetService(r.Context(), cluster, service)
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte(err.Error()))
+		handleError(w, err)
 		return
 	}
 
-	if len(output.Services) == 0 {
-		w.WriteHeader(http.StatusNotFound)
-		w.Write([]byte("Not Found"))
-		return
-	}
-
-	events := output.Services[0].Events
+	events := output.Events
 	j, err := json.Marshal(events)
 	if err != nil {
-		log.Errorf("cannot marshal response (%v) into JSON: %s", events, err)
-		w.WriteHeader(http.StatusInternalServerError)
+		handleError(w, err)
 		return
 	}
 
@@ -517,20 +428,18 @@ func (s *server) ServiceLogsHandler(w http.ResponseWriter, r *http.Request) {
 	logStream := fmt.Sprintf("%s/%s/%s", service, container, task)
 	log.Debugf("getting events for log group/stream: %s/%s", cluster, logStream)
 
-	output, err := logService.Service.GetLogEventsWithContext(r.Context(), &cloudwatchlogs.GetLogEventsInput{
+	output, err := logService.GetLogEvents(r.Context(), &cloudwatchlogs.GetLogEventsInput{
 		LogGroupName:  aws.String(cluster),
 		LogStreamName: aws.String(logStream),
 	})
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte(err.Error()))
+		handleError(w, err)
 		return
 	}
 
 	j, err := json.Marshal(output)
 	if err != nil {
-		log.Errorf("cannot marshal response (%v) into JSON: %s", output, err)
-		w.WriteHeader(http.StatusInternalServerError)
+		handleError(w, err)
 		return
 	}
