+++ cloudwatchlogs/cloudwatchlogs.go
@@ -1,18 +1,22 @@
 package cloudwatchlogs
 
 import (
+	"context"
+
+	"github.com/YaleSpinup/ecs-api/apierror"
 	"github.com/YaleSpinup/ecs-api/common"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/cloudwatchlogs"
+	"github.com/aws/aws-sdk-go/service/cloudwatchlogs/cloudwatchlogsiface"
 	log "github.com/sirupsen/logrus"
 )
 
 // CloudWatchLogs is the internal cloudwatch logsobject which holds session
 // and configuration information
 type CloudWatchLogs struct {
-	Service *cloudwatchlogs.CloudWatchLogs
+	Service cloudwatchlogsiface.CloudWatchLogsAPI
 }
 
 // NewSession builds a new aws cloudwatchlogs session
@@ -25,4 +29,18 @@ func NewSession(account common.Account) CloudWatchLogs {
 	}))
 	c.Service = cloudwatchlogs.New(sess)
 	return c
+}
+
+func (c *CloudWatchLogs) GetLogEvents(ctx context.Context, input *cloudwatchlogs.GetLogEventsInput) (*cloudwatchlogs.GetLogEventsOutput, error) {
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	output, err := c.Service.GetLogEventsWithContext(ctx, input)
+	if err != nil {
+		// TODO: return ErrCode("failed to get log events", err)
+		return nil, err
+	}
+
+	return output, nil
 }
