+++ orchestration/cluster.go
@@ -3,14 +3,10 @@ package orchestration
 import (
 	"context"
 	"errors"
-	"fmt"
-	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 
 	"github.com/aws/aws-sdk-go/service/ecs"
-	"github.com/aws/aws-sdk-go/service/ecs/ecsiface"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -18,11 +14,11 @@ import (
 // it will be used, otherwise if the ClusterName is given, it will be created.  If neither is provided, an error
 // will be returned.
 func (o *Orchestrator) processCluster(ctx context.Context, input *ServiceOrchestrationInput) (*ecs.Cluster, error) {
-	client := o.ECS.Service
+	client := o.ECS
 	if input.Service.Cluster != nil {
 		log.Infof("Using provided cluster name (input.Service.Cluster) %s", aws.StringValue(input.Service.Cluster))
 
-		cluster, err := getCluster(ctx, client, input.Service.Cluster)
+		cluster, err := client.GetCluster(ctx, input.Service.Cluster)
 		if err != nil {
 			return nil, err
 		}
@@ -69,7 +65,7 @@ func (o *Orchestrator) processCluster(ctx context.Context, input *ServiceOrchest
 			}
 		}
 
-		cluster, err := createCluster(ctx, client, input.Cluster)
+		cluster, err := client.CreateCluster(ctx, input.Cluster)
 		if err != nil {
 			return nil, err
 		}
@@ -78,105 +74,4 @@ func (o *Orchestrator) processCluster(ctx context.Context, input *ServiceOrchest
 		return cluster, nil
 	}
 	return nil, errors.New("A new or existing cluster is required")
-}
-
-// createCluster creates a cluster with context and name
-func createCluster(ctx context.Context, client ecsiface.ECSAPI, cluster *ecs.CreateClusterInput) (*ecs.Cluster, error) {
-	log.Debugf("creating cluster with input %+v", cluster)
-
-	output, err := client.CreateClusterWithContext(ctx, cluster)
-	if err != nil {
-		return nil, err
-	}
-	return output.Cluster, err
-}
-
-// getCluster gets the details of a cluster with context by the cluster name
-func getCluster(ctx context.Context, client ecsiface.ECSAPI, name *string) (*ecs.Cluster, error) {
-	output, err := client.DescribeClustersWithContext(ctx, &ecs.DescribeClustersInput{
-		Clusters: []*string{name},
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	if len(output.Failures) > 0 {
-		log.Warnf("describe clusters %s returned failures %+v", aws.StringValue(name), output.Failures)
-	}
-
-	if len(output.Clusters) == 0 {
-		msg := fmt.Sprintf("cluster %s not found", aws.StringValue(name))
-		return nil, errors.New(msg)
-	} else if len(output.Clusters) > 1 {
-		return nil, errors.New("unexpected number of clusters returned")
-	}
-
-	return output.Clusters[0], err
-}
-
-// deleteCluster deletes a(n empty) cluster
-func deleteCluster(ctx context.Context, client ecsiface.ECSAPI, name *string) error {
-	_, err := client.DeleteClusterWithContext(ctx, &ecs.DeleteClusterInput{Cluster: name})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// deleteClusterWithRetry continues to retry deleting a cluster until the context is cancelled or it succeeds
-func deleteClusterWithRetry(ctx context.Context, client ecsiface.ECSAPI, arn *string) chan string {
-	cluChan := make(chan string, 1)
-	go func() {
-		t := 1 * time.Second
-		for {
-			if ctx.Err() != nil {
-				log.Debug("cluster delete context is cancelled")
-				return
-			}
-
-			cluster, err := getCluster(ctx, client, arn)
-			if err != nil {
-				log.Errorf("error finding cluster to delete %s: %s", aws.StringValue(arn), err)
-				cluChan <- "unknown"
-				return
-			}
-			log.Debugf("found cluster %+v", cluster)
-
-			t *= 2
-			c := aws.Int64Value(cluster.RegisteredContainerInstancesCount)
-			if c > 0 {
-				log.Infof("found cluster %s, but registered instance count is > 0 (%d)", aws.StringValue(cluster.ClusterName), c)
-				time.Sleep(t)
-				continue
-			} else {
-				log.Infof("found cluster %s with registered instance count of 0, attempting to delete", aws.StringValue(cluster.ClusterName))
-				err := deleteCluster(ctx, client, arn)
-				if err != nil {
-					if awsErr, ok := err.(awserr.Error); ok {
-						switch aerr := awsErr.Code(); aerr {
-						case ecs.ErrCodeClusterContainsContainerInstancesException,
-							ecs.ErrCodeClusterContainsServicesException,
-							ecs.ErrCodeClusterContainsTasksException,
-							ecs.ErrCodeLimitExceededException,
-							ecs.ErrCodeResourceInUseException,
-							ecs.ErrCodeServerException:
-							log.Warnf("unable to remove cluster %s: %s", aws.StringValue(arn), err)
-							time.Sleep(t)
-							continue
-						default:
-							log.Errorf("failed removing cluster %s: %s", aws.StringValue(arn), err)
-							cluChan <- "failure"
-							return
-						}
-					}
-				}
-			}
-
-			cluChan <- "success"
-			return
-		}
-	}()
-
-	return cluChan
 }
